Rename handlers/index.ts to handlers/handlers.ts for clarity
